Updated to get the host element rather than the modal portal itself
