+++ packages/wonder-blocks-modal/index.js
@@ -3,12 +3,12 @@ import ModalLauncher from "./components/modal-launcher.js";
 import StandardModal from "./components/standard-modal.js";
 import TwoColumnModal from "./components/two-column-modal.js";
 import OneColumnModal from "./components/one-column-modal.js";
-import getNearestModalLauncherPortal from "./util/get-nearest-modal-launcher-portal.js";
+import maybeGetPortalMountedModalHostElement from "./util/maybe-get-portal-mounted-modal-host-element.js";
 
 export {
     ModalLauncher,
     StandardModal,
     TwoColumnModal,
     OneColumnModal,
-    getNearestModalLauncherPortal,
+    maybeGetPortalMountedModalHostElement,
 };
